drivers device merge nuttx and posix Device and CDev
