+++ src/drivers/device/device.h
@@ -33,18 +33,5 @@
 
 #pragma once
 
-#ifdef __PX4_NUTTX
-#include "nuttx/device_nuttx.h"
-
-#include <nuttx/arch.h>
-#define ATOMIC_ENTER irqstate_t flags = px4_enter_critical_section()
-#define ATOMIC_LEAVE px4_leave_critical_section(flags)
-
-#elif defined (__PX4_POSIX)
-#include "posix/vdev.h"
-
-#define ATOMIC_ENTER lock()
-#define ATOMIC_LEAVE unlock()
-
-#endif
-
+#include "Device.hpp"
+#include "CDev.hpp"
