+++ src/drivers/device/nuttx/cdev_platform.hpp
@@ -1,0 +1,18 @@
+
+#pragma once
+
+#include <cinttypes>
+
+#include <px4_micro_hal.h>
+#include <nuttx/arch.h>
+
+#define ATOMIC_ENTER irqstate_t flags = px4_enter_critical_section()
+#define ATOMIC_LEAVE px4_leave_critical_section(flags)
+
+namespace device
+{
+
+using px4_file_operations_t = struct file_operations;
+using file_t = struct file;
+
+} // namespace device
