+++ src/drivers/device/posix/vfile.cpp
@@ -39,9 +39,9 @@
  */
 
 #include "vfile.h"
-#include <stdio.h>
 
-using namespace device;
+namespace device
+{
 
 VFile::VFile(const char *fname, mode_t mode) :
 	CDev("vfile", fname)
@@ -51,11 +51,12 @@ VFile::VFile(const char *fname, mode_t mode) :
 VFile *VFile::createFile(const char *fname, mode_t mode)
 {
 	VFile *me = new VFile(fname, mode);
-	me->register_driver(fname, me);
+	px4_file_operations_t *fops = nullptr;
+	register_driver(fname, fops, 0666, (void *)me);
 	return me;
 }
 
-ssize_t VFile::write(device::file_t *handlep, const char *buffer, size_t buflen)
+ssize_t VFile::write(file_t *handlep, const char *buffer, size_t buflen)
 {
 	// ignore what was written, but let pollers know something was written
 	poll_notify(POLLIN);
@@ -63,3 +64,4 @@ ssize_t VFile::write(device::file_t *handlep, const char *buffer, size_t buflen)
 	return buflen;
 }
 
+} // namespace device
